fix(icon.ts-iconComponent.ts): added dynamic svg id support
Some svg elements references another using url and 'id', which should be guaranteed to be unique among the whold html page. There are circumstances that when transitioning from one page to another, the svg fill, stroke, path and many other element are not rendered correctly if ajax enabled. The solution is to make every svg element and its children use unique id and reference it.

use regexp to implement. May not have supported all the causes. Need more testing.